# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Native Data Platform API",
    "description": "Governed local data platform: authentication and RBAC, multi-engine query gateway, DataHub-backed catalog and lineage, versioned metrics, and checkpointed evidence-bound analyses.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasources": {
      "get": {
        "tags": [
          "datasources"
        ],
        "summary": "List Datasources",
        "operationId": "list_datasources_api_v1_datasources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "datasources"
        ],
        "summary": "Create Datasource",
        "operationId": "create_datasource_api_v1_datasources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasources/{datasource_id}": {
      "patch": {
        "tags": [
          "datasources"
        ],
        "summary": "Patch Datasource",
        "operationId": "patch_datasource_api_v1_datasources__datasource_id__patch",
        "parameters": [
          {
            "name": "datasource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Datasource Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasourcePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasources/{datasource_id}/test": {
      "post": {
        "tags": [
          "datasources"
        ],
        "summary": "Test Datasource",
        "operationId": "test_datasource_api_v1_datasources__datasource_id__test_post",
        "parameters": [
          {
            "name": "datasource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Datasource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasources/{datasource_id}/sync": {
      "post": {
        "tags": [
          "datasources"
        ],
        "summary": "Sync Datasource",
        "description": "Queue a DataHub ingestion for the datasource's registered datasets.",
        "operationId": "sync_datasource_api_v1_datasources__datasource_id__sync_post",
        "parameters": [
          {
            "name": "datasource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Datasource Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestions/{ingestion_id}": {
      "get": {
        "tags": [
          "datasources"
        ],
        "summary": "Get Ingestion",
        "operationId": "get_ingestion_api_v1_ingestions__ingestion_id__get",
        "parameters": [
          {
            "name": "ingestion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ingestion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset",
        "operationId": "get_dataset_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/schema": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Schema",
        "operationId": "get_dataset_schema_api_v1_datasets__dataset_id__schema_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/lineage": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Lineage",
        "operationId": "get_dataset_lineage_api_v1_datasets__dataset_id__lineage_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetSchemaResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "schema_hash": {
            "type": "string",
            "title": "Schema Hash"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnInfo"
            },
            "type": "array",
            "title": "Columns"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "source": {
            "type": "string",
            "enum": [
              "snapshot",
              "introspection"
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "schema_hash",
          "columns",
          "fetched_at",
          "expires_at",
          "source"
        ],
        "title": "DatasetSchemaResponse"
      },
...
      "DatasourceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "postgres",
              "mysql",
              "doris"
            ],
            "title": "Kind"
          },
          "connection_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostgresConnectionConfig"
              },
              {
                "$ref": "#/components/schemas/MySQLConnectionConfig"
              },
              {
                "$ref": "#/components/schemas/DorisConnectionConfig"
              }
            ],
            "title": "Connection Config"
          },
          "secret_ref": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Secret Ref"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "kind",
          "connection_config",
          "secret_ref"
        ],
        "title": "DatasourceCreate"
      },
      "DatasourceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasourceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DatasourceListResponse"
      },
      "DatasourcePatch": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9._-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "connection_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostgresConnectionConfig"
              },
              {
                "$ref": "#/components/schemas/MySQLConnectionConfig"
              },
              {
                "$ref": "#/components/schemas/DorisConnectionConfig"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Config"
          },
          "secret_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 3,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Ref"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "DatasourcePatch"
      },
      "DatasourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "connection_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Connection Config"
          },
          "secret_ref": {
            "type": "string",
            "title": "Secret Ref"
          },
          "capabilities": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "health_status": {
            "type": "string",
            "title": "Health Status"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "connection_config",
          "secret_ref",
          "capabilities",
          "enabled",
          "health_status",
          "created_by",
          "created_at",
          "version"
        ],
        "title": "DatasourceResponse"
      },
      "DatasourceTestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "server_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Version"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "capabilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderCapabilitiesDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "server_version",
          "latency_ms",
          "checked_at"
        ],
        "title": "DatasourceTestResponse"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestionTaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "datasource_id": {
            "type": "string",
            "format": "uuid",
            "title": "Datasource Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "datasource_id",
          "status",
          "created_at"
        ],
        "title": "IngestionTaskResponse"
      },
...
      "LineageResponse": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/LineageNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Edges"
          },
          "filtered_nodes": {
            "type": "integer",
            "title": "Filtered Nodes",
            "default": 0
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "analysis_evidence": {
            "items": {
              "$ref": "#/components/schemas/AnalysisEvidenceEdge"
            },
            "type": "array",
            "title": "Analysis Evidence"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "direction",
          "depth",
          "status"
        ],
        "title": "LineageResponse"
      },
...
    }
  },
...
}